move v3_upgrade_helper to top leve
